General cleanup and new process skeleton script
Given the new approach to use small scripts to accomplish discrete steps vs a monolithic processing script the run.py has been renamed to convert.py and simplified to only convert .hex files to _converted.csv files.
Along those same lines a new process.py file have been created.  This file takes an ini file and an optional converted data file for processing.

So the new workflow for processing data is a 2-step process

convert.py -o ./GS3601101 ./GS3601101.hex ./GS3601101.XMLCON

process.py -d -i ./GS3601101/GS3601101_converted.csv ./configuration.ini

Using this approach it should be possible to batch process casts using a custom ini for each cast or using a common ini for all casts.

Optionally the input file can be specified in the ini file by specifying an 'input_data_file' varable in the 'ctd_processing' section of the ini file.
This would remove the need to specify an input file (via the '-i [inputFile]' argument)

Other features of process.py include:
Setting an output directory in the command line for any output files (via '-o outputDir')
Setting an output directory in the ini file ('output_data_directory' varable in the 'ctd_processing' section)
If an input file is specified the output directory is set to the same directory as the input.
If no input file or output directory are specified the default output directory is set to the current directory
Specifying an output directory overrides all other options for the output directory.
